update shop query
