+++ NineChronicles.DataProvider/GraphQL/Types/NineChroniclesSummaryQuery.cs
@@ -1,6 +1,5 @@
 ﻿namespace NineChronicles.DataProvider.GraphQL.Types
 {
-    using System.Linq;
     using global::GraphQL;
     using global::GraphQL.Types;
     using Libplanet;
@@ -54,86 +53,10 @@
                         : new Address(address.Replace("0x", string.Empty));
                     return Store.GetAvatars(avatarAddress);
                 });
-            Field<IntGraphType>(
-                name: "ShopEquipmentCount",
-                resolve: context =>
-                {
-                    var shopEquipmentCount = Store.GetShopEquipmentCount();
-                    return shopEquipmentCount;
-                });
-            Field<ListGraphType<ShopEquipmentType>>(
-                name: "ShopEquipments",
-                arguments: new QueryArguments(
-                    new QueryArgument<StringGraphType> { Name = "sellerAvatarAddress" }
-                ),
-                resolve: context =>
-                {
-                    string? address = context.GetArgument<string?>("sellerAvatarAddress", null);
-                    Address? sellerAvatarAddress = address == null
-                        ? (Address?)null
-                        : new Address(address.Replace("0x", string.Empty));
-                    return Store.GetShopEquipments(sellerAvatarAddress);
-                });
-            Field<IntGraphType>(
-                name: "ShopConsumableCount",
-                resolve: context =>
-                {
-                    var shopConsumableCount = Store.GetShopConsumableCount();
-                    return shopConsumableCount;
-                });
-            Field<ListGraphType<ShopConsumableType>>(
-                name: "ShopConsumables",
-                arguments: new QueryArguments(
-                    new QueryArgument<StringGraphType> { Name = "sellerAvatarAddress" }
-                ),
-                resolve: context =>
-                {
-                    string? address = context.GetArgument<string?>("sellerAvatarAddress", null);
-                    Address? sellerAvatarAddress = address == null
-                        ? (Address?)null
-                        : new Address(address.Replace("0x", string.Empty));
-                    return Store.GetShopConsumables(sellerAvatarAddress);
-                });
-            Field<IntGraphType>(
-                name: "ShopCostumeCount",
-                resolve: context =>
-                {
-                    var shopCostumeCount = Store.GetShopCostumeCount();
-                    return shopCostumeCount;
-                });
-            Field<ListGraphType<ShopCostumeType>>(
-                name: "ShopCostumes",
-                arguments: new QueryArguments(
-                    new QueryArgument<StringGraphType> { Name = "sellerAvatarAddress" }
-                ),
-                resolve: context =>
-                {
-                    string? address = context.GetArgument<string?>("sellerAvatarAddress", null);
-                    Address? sellerAvatarAddress = address == null
-                        ? (Address?)null
-                        : new Address(address.Replace("0x", string.Empty));
-                    return Store.GetShopCostumes(sellerAvatarAddress);
-                });
-            Field<IntGraphType>(
-                name: "ShopMaterialCount",
-                resolve: context =>
-                {
-                    var shopMaterialCount = Store.GetShopMaterialCount();
-                    return shopMaterialCount;
-                });
-            Field<ListGraphType<ShopMaterialType>>(
-                name: "ShopMaterials",
-                arguments: new QueryArguments(
-                    new QueryArgument<StringGraphType> { Name = "sellerAvatarAddress" }
-                ),
-                resolve: context =>
-                {
-                    string? address = context.GetArgument<string?>("sellerAvatarAddress", null);
-                    Address? sellerAvatarAddress = address == null
-                        ? (Address?)null
-                        : new Address(address.Replace("0x", string.Empty));
-                    return Store.GetShopMaterials(sellerAvatarAddress);
-                });
+            Field<NonNullGraphType<ShopQuery>>(
+                name: "shopQuery",
+                resolve: context => new ShopQuery(store)
+            );
             Field<ListGraphType<BattleArenaRankingType>>(
                 name: "BattleArenaRanking",
                 arguments: new QueryArguments(
